src/main/java/by/andd3dfx/helper/DateTimeHelper.java
package by.andd3dfx.helper;

import java.time.Instant;
import java.time.LocalDateTime;
import java.time.ZoneOffset;
import java.time.format.DateTimeFormatter;

public class DateTimeHelper {

    public final static DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss'Z'");

    public static LocalDateTime convertInstantToLocalDateTime(Instant instant) {
        return LocalDateTime.ofInstant(instant, ZoneOffset.UTC);
    }

    public static LocalDateTime convertZStringToLocalDateTime(String zString) {
        return convertInstantToLocalDateTime(Instant.parse(zString));
    }
}
